Don't fall back to QGuiApplication for line control's layoutDirection
If the widget whose line control we are has its own locale setting,
then we should use that rather than the default UI layout direction.
Instead, the client of the QWidgetLineControl needs to handle the case
where auto-detection of the text direction can't be done based on the
content of the line control.

Fortunately, QWidget's handling of text direction handles setting to
auto gracefully as unsetting, on which it consults the parent widget
to set a sensible default.

Change-Id: Id0d247f1e6e7219b6d0ece5856ca93ee87778c74
Fixes: QTBUG-53110
Reviewed-by: Orgad Shaneh <orgads@gmail.com>
(cherry picked from commit 589e34eb3210d8725d68941dbeb096424dda68df)